zcu102-arm: Connect separate chip selects to the QSPI flashes
When using the QSPI striping functionality both buses and chip selects are
enabled but communication should only be with the flash connected to the lower
cs on bus 0 and the flash connected to upper cs on bus 1. Sharing the chip
select lines (as real hw) made the QSPI model enable both flashes on both buses
when using the striping functionality.

Signed-off-by: Francisco Iglesias <frasse.iglesias@gmail.com>
[ Changes by AF:
 - Rebase on our latest tree
]
Reviewed-by: Alistair Francis <alistair.francis@xilinx.com>