board-versal-ps-virt: Attach 8Mbit Flash to Cadence SPI
The attachment matches the prebuilt system.dtb in 2019.1 BSP
xilinx-versal-virt-2019.1

The device-images must be specified with the '-blockdev' QEMU option,
e.g.,:
  -blockdev node-name=spi0_flash0,driver=file,filename=flash0.bin
  -blockdev node-name=spi1_flash0,driver=file,filename=flash1.bin

If an option is not given when booting prebuilt Linux on the prebuilt
BSP, Linux dmesg will display warning of the corresponding flash
device as 'read-only'.

To avoid the warning, make sure the hw-dtb and system.dtb match.

Signed-off-by: Tong Ho <tong.ho@xilinx.com>
Signed-off-by: Edgar E. Iglesias <edgar.iglesias@xilinx.com>