versal-rams: ocm: Use MM_OCM_MEM_SIZE macro to define OCM banks
Using the MM_OCM_MEM_SIZE macro allows more flexibiliy if OCM size changes in
the future.

This change also removes the base address of each bank in the device tree's
unit address, where the unit address is the portion of the node that says:
label: node-name@unit-address { ... };

We remove the unit address because if we try to use the MM_OCM_MEM_SIZE macro
to define the unit address, it expands to:
ocm_mem_bank_x@((MM_OCM_MEM_SIZE / 4) * x) { ... };
which is not valid syntax.

Removing the unit address only changes the node's name in QEMU and has no other
effects, however it does break the DTS specification, which says the
unit address must match the first value of the reg property, if present.

Signed-off-by: Joe Komlodi <joe.komlodi@xilinx.com>
Message-Id: <1626716174-130574-2-git-send-email-joe.komlodi@xilinx.com>
Signed-off-by: Edgar E. Iglesias <edgar.iglesias@xilinx.com>