ksb: memory-map: Define the full 2G of low DDR
The RPU has the OCM overlayed at 0x7ff00000 since it's not
able to boot from addresses higher than 0x80000000. The APU
and other bus masters don't see this overlay and have access
to the full DDR.

Define MM_TOP_DDR_SIZE to the full size so we create the
full region. Later we'll implement the RPU OCM overlay for
the RPU only.

Signed-off-by: Edgar E. Iglesias <edgar.iglesias@xilinx.com>