zynqmp: Fix DDR aliasing onto qemu_sysmem
We alias DDR onto qemu_sysmem to allow DMA models that are not
multi address-space aware to still work (with limitations).
Due to size-cells inheritance we were only mapping the lower
2G on to sysmem.

Fix this by moving the alias node into a sub-node of sysmem
now inheriting the appropriate cells.

This fixes an issue with SATA on Linux.

Signed-off-by: Edgar E. Iglesias <edgar.iglesias@xilinx.com>