ultra96-arm: implement the sw4 switch
According to:
https://www.96boards.org/documentation/consumer/ultra96/ultra96-v1/\
hardware-docs/files/ultra96-schematics.pdf

The switch 4 (SW4) is connected to MIO23_GPIO_PB with a pull-up resistor.
Forgetting to model that switch is equivalent to have the button pressed, which
reboot the board during the boot.

This uses the "gpio-key" model from QEMU in order to model SW4.

Signed-off-by: Frederic Konrad <fkonrad@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>