versal-psx-apu: fix the number of IRQs
Versal-Net is supposed to have 256 SPIs, but according to the ARM GIC
documentation [1], the ITLinesNumber field indicates the maximum number of SPIs
that the GIC *might* support.  This also determines the number of implemented
interrupt registers, ie: GICD_IGROUPRn, GICD_ISENABLERn, etc.  Since those
registers are present on real hardware, and that the ITLinesNumber field is 0x10
on real hardware, fix num-irqs to get the same behavior.

[1] https://developer.arm.com/documentation/ihi0048/b/Programmers--Model/\
Distributor-register-descriptions/Interrupt-Controller-Type-Register--GICD-TYPER

Signed-off-by: Frederic Konrad <fkonrad@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>