interconnect.dtsh: introduce MEMATTR_NAME and MEMATTR_SMID_NAME
When declaring a QEMU memory transaction attribute with MEMATTR (or
MEMATTR_SMID), it is implicitly named by concatenating the name given as
the macro parameter with a suffix. Add the two macros MEMATTR_NAME and
MEMATTR_SMID_NAME to retrieve those names to avoid having to know the
generated name.

E.g., instead of doing:

    MEMATTR(foo, NON_SECURE, 0x0);
    [...]
    bar {
        memattr = <&foo_memattr>;
    };

you can now do:

    MEMATTR(foo, NON_SECURE, 0x0);
    [...]
    bar {
        memattr = <&MEMATTR_NAME(foo)>;
    };

Signed-off-by: Luc Michel <luc.michel@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>