versal-ps*-virt: drop apu_intc_redirect
That "zynqmp-intc-redirect" models lay between the GIC output and CPU inputs.
It raises a "pmu-out" GPIO when an IRQ should go to the CPU and that CPU is
powered down.  It is not wired in Versal-Net so it can be dropped.  For Versal
it's replaced with the GIC wakerequest signal, so it can be dropped as well.

Signed-off-by: Frederic Konrad <fkonrad@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>