nxp,pca9544: remove the chip-enable property
This chip-enable property was very misleading as it actually referred to
the 3 LSBs of the I²C chip address (corresponding to the 3 pins found on the
actual chip).

This property has been deprecated in QEMU and the model now relies on the
reg value for its address. Remove those properties and fix the reg value
where needed.

Signed-off-by: Luc Michel <luc.michel@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>