versal-vitis.dtsi: don't drop the ddr node unconditionnally
Comparing xcvp1902-ps-cosim-vitis-virt and xcvp1902-ps-virt device tree output,
it looks like the ddr_2 node is missing which breaks the kernel.

The ddr_2 node is supposedly deleted in order to be added back if needed in
ddr_memspec.dtsi.  Thus if the dtsi doesn't exist it just disappear.  By default
better keep them when ddr_memspec.dtsi isn't included to be in sync with the
non cosim device tree.

Signed-off-by: Frederic Konrad <fkonrad@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>