versal-icnt: declare constants for the XPPU apertures
Declare some constants for the XPPU apertures. Now that the QEMU model
is parametrized using those constants, give them a name so that they can
be overriden for other chips than versal. Also remove the region
property since it is now deprecated. Not specifying it indicate to the
model to use above constants for configuration.

Signed-off-by: Luc Michel <luc.michel@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>