versal2-mmi: add the USB3.2 controller
Add the MMI USB controller. This controller is a DWC3. Note that I'm not
sure about the num-intrs and num-slots property values. The former is
based on the fact that I found two IRQs for this peripheral. The latter
is just a guess. Those can be changed latter if needed.

Signed-off-by: Luc Michel <luc.michel@amd.com>
Signed-off-by: Francisco Iglesias <francisco.iglesias@amd.com>